+++ benchmark/matrix/matrix_from_rows.dart
@@ -7,8 +7,8 @@ const numOfColumns = 1000;
 class MatrixFromRowsBenchmark extends BenchmarkBase {
   MatrixFromRowsBenchmark() : super('Matrix initialization (fromRows)');
 
-  final _source = List<Row>.filled(
-      numOfRows, Row.random(numOfColumns, min: -10000, max: 10000, seed: 12));
+  final _source = List<RowMatrix>.filled(numOfRows,
+      RowMatrix.random(numOfColumns, min: -10000, max: 10000, seed: 12));
 
   static void main() {
     MatrixFromRowsBenchmark().report();